requirements.txt
networkx==3.4.2
python-docx==1.1.2
pyinstaller==6.13.0